design 6.0
