+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb.Android/Resources/drawable/ListCheckWhite.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+        android:width="24dp"
+        android:height="24dp"
+        android:viewportWidth="24.0"
+        android:viewportHeight="24.0"
+        android:tint="?attr/colorControlNormal">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M13,10L3,10c-0.55,0 -1,0.45 -1,1s0.45,1 1,1h10c0.55,0 1,-0.45 1,-1s-0.45,-1 -1,-1zM13,6L3,6c-0.55,0 -1,0.45 -1,1s0.45,1 1,1h10c0.55,0 1,-0.45 1,-1s-0.45,-1 -1,-1zM3,16h6c0.55,0 1,-0.45 1,-1s-0.45,-1 -1,-1L3,14c-0.55,0 -1,0.45 -1,1s0.45,1 1,1zM22.21,12.21l0.09,0.09c0.39,0.39 0.39,1.02 0,1.41l-5.58,5.59c-0.39,0.39 -1.02,0.39 -1.41,0l-3.09,-3.09c-0.39,-0.39 -0.39,-1.02 0,-1.41l0.09,-0.09c0.39,-0.39 1.02,-0.39 1.41,0l2.3,2.3 4.78,-4.79c0.38,-0.4 1.02,-0.4 1.41,-0.01z"/>
+</vector>
