+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb.Android/Resources/drawable/TransferLogoWhite.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+        android:width="24dp"
+        android:height="24dp"
+        android:viewportWidth="24.0"
+        android:viewportHeight="24.0"
+        android:tint="?attr/colorControlNormal">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M9.01,14L3,14c-0.55,0 -1,0.45 -1,1s0.45,1 1,1h6.01v1.79c0,0.45 0.54,0.67 0.85,0.35l2.78,-2.79c0.19,-0.2 0.19,-0.51 0,-0.71l-2.78,-2.79c-0.31,-0.32 -0.85,-0.09 -0.85,0.35L9.01,14zM14.99,11.79L14.99,10L21,10c0.55,0 1,-0.45 1,-1s-0.45,-1 -1,-1h-6.01L14.99,6.21c0,-0.45 -0.54,-0.67 -0.85,-0.35l-2.78,2.79c-0.19,0.2 -0.19,0.51 0,0.71l2.78,2.79c0.31,0.31 0.85,0.09 0.85,-0.36z"/>
+</vector>
