+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb/LoginPage.xaml
@@ -13,93 +13,92 @@
                          Padding="50" AnchorY="50">
                 <Grid Padding="5">
                     <Grid.RowDefinitions>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
+                            <RowDefinition Height="Auto"/>
+                            <RowDefinition Height="Auto"/>
+                            <RowDefinition Height="Auto"/>
+                            <RowDefinition Height="Auto"/>
+                            <RowDefinition Height="Auto"/>
+                            <RowDefinition Height="Auto"/>
                     </Grid.RowDefinitions>
                     <Grid.ColumnDefinitions>
-                        <ColumnDefinition Width="Auto"/>
-                        <ColumnDefinition Width="*"/>
-                        <ColumnDefinition Width="Auto"/>
+                            <ColumnDefinition Width="Auto"/>
+                            <ColumnDefinition Width="*"/>
+                            <ColumnDefinition Width="Auto"/>
                     </Grid.ColumnDefinitions>
-
                     <Image x:Name="LogoUlacit" 
-                       HorizontalOptions="Center" 
-                       Source="UlacitLogo.png"
-                       WidthRequest="130"
-                       VerticalOptions="Start"
-                       Grid.Row="0"
-                       Grid.Column="1"
-                       Grid.ColumnSpan="3" 
-                       ></Image>
+                           HorizontalOptions="Center" 
+                           Source="UlacitLogo.png"
+                           WidthRequest="130"
+                           VerticalOptions="Start"
+                           Grid.Row="0"
+                           Grid.Column="1"
+                           Grid.ColumnSpan="3" 
+                           ></Image>
                     <Label Text="Banco ULACIT" 
-					   Margin="0, 10, 0, 50"
-                       HorizontalTextAlignment="Center"
-                       HorizontalOptions="Center"
-					   TextColor="White" 
-					   FontSize="25" 
-					   FontAttributes="Bold" 
-                       Grid.Row="1"
-                       Grid.Column="1"
-                       Grid.ColumnSpan="3"></Label>
+				           Margin="0, 10, 0, 50"
+                           HorizontalTextAlignment="Center"
+                           HorizontalOptions="Center"
+					       TextColor="White" 
+		        	       FontSize="25" 
+			    	       FontAttributes="Bold" 
+                           Grid.Row="1"
+                           Grid.Column="1"
+                           Grid.ColumnSpan="3"></Label>
                     <Image Source="LoginWhite.xml" 
-                       Grid.Row="2"
-                       Grid.Column="1"
-                       Grid.ColumnSpan="3"
-                       HorizontalOptions="Start"></Image>
+                           Grid.Row="2"
+                           Grid.Column="1"
+                           Grid.ColumnSpan="3"
+                           HorizontalOptions="Start"></Image>
                     <Entry x:Name="Txt_Username" 
-                       Grid.Row="2"
-                       Grid.Column="1"
-                       Grid.ColumnSpan="3"
-                       Placeholder="Usuario"
-					   PlaceholderColor="#FFFFFF"
-                       Text="#FFFFFF"
-                       HorizontalTextAlignment="Center"
-					   TextColor="#C5C2C2" 
-                       Margin="0,0,0,0"></Entry>
+                           Grid.Row="2"
+                           Grid.Column="1"
+                           Grid.ColumnSpan="3"
+                           Placeholder="Usuario"
+					       PlaceholderColor="#FFFFFF"
+                           Text="#FFFFFF"
+                           HorizontalTextAlignment="Center"
+					       TextColor="#C5C2C2" 
+                           Margin="0,0,0,0"></Entry>
                     <Image Source="LockWhite.xml" 
                            Grid.Row="3"
                            Grid.Column="1"
                            Grid.ColumnSpan="3"
                            HorizontalOptions="Start"></Image>
                     <Entry x:Name="Txt_Password"
-                       Grid.Row="3"
-                       Grid.Column="1"
-                       Grid.ColumnSpan="3"
-                       IsPassword="True"
-                       Placeholder="Contraseña"
-					   Margin="0, 0, 0, 0"
-                       PlaceholderColor="#FFFFFF"
-                       Text="#FFFFFF"
-                       HorizontalTextAlignment="Center"
-					   TextColor="#C5C2C2" ></Entry>
+                           Grid.Row="3"
+                           Grid.Column="1"
+                           Grid.ColumnSpan="3"
+                           IsPassword="True"
+                           Placeholder="Contraseña"
+					       Margin="0, 0, 0, 0"
+                           PlaceholderColor="#FFFFFF"
+                           Text="#FFFFFF"
+                           HorizontalTextAlignment="Center"
+					       TextColor="#C5C2C2" ></Entry>
                     <Button x:Name="Btn_Login" 
-                        Grid.Row="4"
-                        Grid.Column="1"
-                        Grid.ColumnSpan="3"
-                        Text="Iniciar Sesion" 
-						Margin="0, 25, 0, 0"
-                        BorderColor="#252525" 
-						BackgroundColor="#EA8B3C" 
-                        TextColor="#000000"
-						CornerRadius="30" 
-                        Clicked="Btn_Login_Clicked" FontAttributes="Bold"
-                        FontFamily="FontAwesome.ot#FontAwesome">
-                </Button>
-                <Button x:Name="Btn_Register" 
-                        Grid.Row="5"
-                        Grid.Column="1"
-                        Grid.ColumnSpan="3"
-                        Text="¿No tiene una Cuenta?" 
-                        TextColor="White"
-						BorderColor="Transparent"
-						BackgroundColor="Transparent"
-                        CornerRadius="10" 
-                        Clicked="Btn_Register_Clicked" 
-                        FontAttributes="Bold"></Button>
+                           Grid.Row="4"
+                           Grid.Column="1"
+                           Grid.ColumnSpan="3"
+                           Text="Iniciar Sesion" 
+						   Margin="0, 25, 0, 0"
+                           BorderColor="#252525" 
+        				   BackgroundColor="#EA8B3C" 
+                           TextColor="#000000"
+						   CornerRadius="30" 
+                           Clicked="Btn_Login_Clicked" FontAttributes="Bold"
+                           FontFamily="FontAwesome.ot#FontAwesome">
+                    </Button>
+                    <Button x:Name="Btn_Register" 
+                           Grid.Row="5"
+                           Grid.Column="1"
+                           Grid.ColumnSpan="3"
+                           Text="¿No tiene una Cuenta?" 
+                           TextColor="White"
+						   BorderColor="Transparent"
+						   BackgroundColor="Transparent"
+                           CornerRadius="10" 
+                           Clicked="Btn_Register_Clicked" 
+                           FontAttributes="Bold"></Button>
                 <!--<Button x:Name="Btn_Agregar" 
                         Text="Agregar Usuario" 
                         BorderColor="Navy"
