+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb/Models/Transferencia.cs
@@ -13,11 +13,25 @@ namespace AppBancaEnLineaWeb.Models
         public string TRA_DESCRIPCION { get; set; }
         public char TRA_ESTADO { get; set; }
         public DateTime TRA_FECHA { get; set; }
-        public decimal TRA_MONTO { get; set; }
-        #endregion
+        public decimal TRA_MONTO_ORIGEN { get; set; }
+        public decimal TRA_MONTO_FINAL { get; set; }
+        public string PAG_DESCRIPCION_label
+        {
+            get
+            {
+                return string.Format("Transferencia #{0} | Cuenta Origen #{1} | Cuenta Destino: {2} | Monto: {3} | Fecha: {4}",
+                    TRA_CODIGO
+                    , CUE_ORIGEN
+                    , CUE_DSTINO
+                    , TRA_MONTO_ORIGEN
+                    , TRA_MONTO_FINAL
+                    , TRA_FECHA.ToString("dd/MM/yyyy  hh:mm"));
+            }
+        }
+#endregion
 
-        #region Constructor
-        public Transferencia()
+#region Constructor
+public Transferencia()
         {
         }
         #endregion
