+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb/Views/PagoPageList.xaml
@@ -3,8 +3,8 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              x:Class="AppBancaEnLineaWeb.Views.PagoPageList"
              BackgroundColor="#FFFFFF"
-             Title="Lista"
-             Icon="round_account_balance_24.xml">
+             Title="Pagos Realizados"
+             Icon="ListCheckWhite.xml">
     <ContentPage.Content>
         <StackLayout Padding="25">
             <Label Text="Pagos Realizados" 
@@ -12,7 +12,6 @@
                 VerticalOptions="StartAndExpand"
 				TextColor="#000000" 
 				FontSize="26"
-				TextDecorations="Underline"
 				FontAttributes="Bold" Margin="0,20,0,0">
             </Label>
             <Label
@@ -23,25 +22,27 @@
                 <ListView x:Name="PagoList" BackgroundColor="#FFFFFF" Grid.Row="4" SeparatorVisibility="Default">
                 <ListView.ItemTemplate>
                     <DataTemplate>
-                        <ViewCell Height="30">
+                        <ViewCell Height="35">
                             <StackLayout Padding="5">
                                 <Label 
                                     Text="{Binding PAG_DESCRIPCION_label}"
                                     TextColor="#000000"
-                                    FontSize="Medium"></Label>
+                                    FontSize="Small"></Label>
                             </StackLayout>
                         </ViewCell>
                     </DataTemplate>
                 </ListView.ItemTemplate>
             </ListView>
             <Button 
-					Image="actualizar.png"
-                    Padding="15"
+					Image="RefreshBlack.xml"
+                    FontSize="Large"
+                    WidthRequest="30"
+                    HeightRequest="30"
                     BackgroundColor="Beige"  
-                    CornerRadius="50"
+                    CornerRadius="100"
 					HorizontalOptions="End"
                     Clicked="Btn_RefrescarPantalla"
-                    Margin="15, 0"/>
+                    Margin="0, 10, 0, 5"/>
         </StackLayout>
     </ContentPage.Content>
 </ContentPage>
