+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb/Views/TransferenciaPage.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              x:Class="AppBancaEnLineaWeb.Views.TransferenciaPage"
              Title="Transferencias"
-             Icon="PaymentsWhite.xml"
+             Icon="TransferLogoWhite.xml"
              BackgroundColor="#FFFFFF">
     <ContentPage.Content>
         <StackLayout Padding="30" >
@@ -13,7 +13,6 @@
 				TextColor="#000000"
                 VerticalOptions="StartAndExpand"
 				FontSize="Large"
-				TextDecorations="Underline"
 				FontAttributes="Bold" 
                 Margin="0,10,0,0">
             </Label>
