+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb/Views/TransferenciaPageList.xaml.cs
@@ -3,23 +3,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Transactions;
 
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
+
+using AppBancaEnLineaWeb.Controllers;
+using AppBancaEnLineaWeb.Models;
+using AppBancaEnLineaWeb.Views;
 
 namespace AppBancaEnLineaWeb.Views
 {
 	[XamlCompilation(XamlCompilationOptions.Compile)]
 	public partial class TransferenciaPageList : ContentPage
 	{
-		public TransferenciaPageList ()
-		{
-			InitializeComponent ();
-		}
-
-        private void Btn_Volver(object sender, EventArgs e)
-        {
-            App.Current.MainPage = new MainTabbedPage();
-        }
+        
     }
 }
