+++ AppBancaEnLineaWeb/AppBancaEnLineaWeb/AppBancaEnLineaWeb/Views/UsuarioPage.xaml
@@ -60,18 +60,14 @@
                 Margin="0,0,0,40"
                 Text="Fecha: ">
             </Label>
-            <Label
-                VerticalOptions="EndAndExpand">
-            </Label>
-            <!--<Button 
-                x:Name="Btn_Actualizar"
-                Text="Actualizar"
-                TextColor="White"
-                CornerRadius="30"
-                BackgroundColor="#61518A"
-                Margin="10, 0"
-                >
-            </Button>-->
+            <Button Grid.Column="0" 
+                    Text="CERRAR SESIÓN"
+                    TextColor="White"
+                    BackgroundColor="DarkRed"
+                    HorizontalOptions="FillAndExpand"
+                    VerticalOptions="EndAndExpand"
+                    CornerRadius="30"
+                    Clicked="CerrarSesionTapped"/>
             
         </StackLayout>
     </ContentPage.Content>
